test(popover): recategorize and clean up browser tests
